refactor: separate SecurityConfig responsibilities
Refatoração da configuração de segurança:
- SecurityConfig mantém apenas HttpSecurity, filtros JWT e OAuth2Login.
- AuthConfig concentra beans de autenticação (BCryptPasswordEncoder, AuthenticationManager, CustomOAuth2UserService).
- CorsConfig concentra a configuração de CORS.
