+++ src/main/java/com/maisprati/hub/infrastructure/security/config/AuthConfig.java
@@ -1,0 +1,32 @@
+package com.maisprati.hub.infrastructure.security.config;
+
+import com.maisprati.hub.infrastructure.persistence.repository.UserRepository;
+import com.maisprati.hub.infrastructure.security.oauth2.CustomOAuth2UserService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+
+@Configuration
+@RequiredArgsConstructor
+public class AuthConfig {
+	
+	private final UserRepository userRepository;
+	
+	@Bean
+	public BCryptPasswordEncoder passwordEncoder() {
+		return new BCryptPasswordEncoder();
+	}
+	
+	@Bean
+	public AuthenticationManager authenticationManager(AuthenticationConfiguration authConfig) throws Exception {
+		return authConfig.getAuthenticationManager();
+	}
+	
+	@Bean
+	public CustomOAuth2UserService customOAuth2UserService() {
+		return new CustomOAuth2UserService(userRepository);
+	}
+}
